requirements.txt
numpy
Pandas
plotly.express
Streamline
